Fix Docker setup
In Docker, when setting an existing folder as a volume it "links" the
host folder with the one inside the container, making the changes made
during build futile.

The application is installed when creating the image not running it, the
reason for this is to increase flexibility for various setups (ex. Kubernetes)